Change TestResult comparison so errors are last
For exhaustive search, I sort TestResults and then select the first
TestResult in the sorted list as the optimal point.  When some points
are in range and some are out of range, the out-of-range points need to
sort *after* the in-range points; otherwise, one of the out-of-range
points will be selected as the optimum.